+++ app.js
@@ -1,131 +1,67 @@
-"use strict";
 
-const koa = require('koa');
-const config = require('./config/config');
-const secret = require('./config/secret');
-const logger = require('koa-logger');
+var express = require('express');
+var http = require('http');
+var path = require('path');
 
-const router = require('koa-router')();
-const koaBody = require('koa-body')();
-
-
-const jade = require('koa-jade');
-const views = require('koa-views');
-const serve = require('koa-static');
-
-const mongoose = require('mongoose');
-const Story = require('./models/Story');
-
-require('node-jsx').install();
-const react = require('react');
-const ArticleNavComponent = react.createFactory( require('./src/js/components/ArticleNav') );
-const StoryComponent = react.createFactory( require('./src/js/components/Story') );
-
-//mongoose.connect( secret.mongodb );
+var morgan = require('morgan');
+var cookieParser = require('cookie-parser');
+var bodyParser = require('body-parser');
+var session = require('express-session');
 
 
-const app = koa();
+var config = require('./config/config');
+var secret = require('./config/secret');
 
-app.use(logger());
-app.use(jade.middleware({
-  viewPath: 'views',
-  debug: true,
-  noCache: true,
-  basedir: 'views'
-}));
 
-app.use(serve('public'));
-app.use(function *(next){
-  yield next;
-});
-app.use(function *(next){
-  var stories = yield Story.find().sort({updatedAt: -1});
-  this.state.stories = stories || [];
-  this.state.latest_story = stories[0];
-  this.state.article_nav = react.renderToString( ArticleNavComponent({stories}) );
-  this.state.brand = config.brand;
-  this.state.title = config.title;
-  this.state.author = config.author;
-  yield next;
-});
+var mongoose = require('mongoose');
+var configDb = require('./config/database')(mongoose);
 
-// response
-router
-  .get('/', function *(next) {
-    if(this.state.stories.length > 0 ){
-      this.state.story = react.renderToString( StoryComponent({story: this.state.latest_story.marked() }) )
-      this.render( 'story' )
-    }else{
-      this.render('add-article');
-    }
+var app = express();
+var router = express.Router();
+var port = process.env.PORT || 2000;
 
-  })
-  .get('/:slug', function *(next){
-    var story = yield Story.findOne( {slug: this.params.slug} );
 
-    if( story ){
-      this.state.title = story.title;
-      this.state.story = react.renderToString( StoryComponent({story: story.marked()}) );
-    }
+// Config express
 
-    this.render( 'story' );
-  })
-  .post('/add-article', koaBody, function *(next){
-    var body = this.request.body;
-    var id = this.request.body.parent_article_id;
+app.use(cookieParser());
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({extended: true}));
 
-    var post, res;
-    if(id){
-      story = yield Story.findOne({_id: id});
-      story.posts.push(body);
-      res = yield story.save();
-    }else{
-      var story = new Story(body);
-      res  = yield story.save();
-    }
-    this.body = res;
-  })
-  .get('/hello', function *(next) {
-    yield this.render('hello')
-  })
-  .get('/about', function *(next) {
-    this.body = "<!DOCTYPE html><head><title>About</title></head><body>About Us</body></html>";
-  })
-  .get('/react', function *(next) {
-    yield this.render('react')
-  })
-  .get('/json', function *(next) {
-    this.body = {
-      name: 'brett',
-      number: 83,
-      isProgrammer: true
-    };
-  })
-  .get('/admin/add-article', function *(next){
-    this.render('add-article');
-  })
-  .get('/admin/edit-article/:story_id', function *(next){
-    var story = yield Story.findOne({_id: this.params.story_id});
-    this.render('add-article',{
-      article: story
-    });
-  })
-  .get('/admin/edit-article/:story_id/:post_id', function *(next){
-    var story = yield Story.findOne({slug: slug});
-    var post = story.posts.id(this.params.id);
-    this.render('add-article',{
-      article: post
-    });
-  });
+app.set('views', path.join(__dirname, 'views'));
+app.use(express.static(path.join(__dirname, 'src')));
+app.set('view engine', 'jade');
 
-app
-  .use(router.routes())
-  .use(router.allowedMethods());
-module.exports = app;
+app.use(session({secret: secret.sessionSecret}));
 
-// start app if it isn't being required into another module
-if (!module.parent) {
-  const port = process.env.PORT || 9999;
-  app.listen(port);
-  console.log('Listening on port ' + port);
+app.listen(port);
+console.log('Express server listening on port '+ port);
+
+
+//Development
+
+//change for deploy
+var env = process.env.NODE_ENV || 'development'; 
+if(env == 'development'){
+  app.use(morgan('dev'));
+  app.locals.pretty = true;
 }
+
+//Routes
+
+var Stories = require('./models/Story');
+
+// route middleware that will happen on every request
+app.use(function(req, res, next) {
+
+    // log each request to the console
+    console.log(req.method, req.url);
+
+    // continue doing what we were doing and go to the route
+    next(); 
+});
+
+app.brand = config.brand;
+app.title = config.title;
+app.author = config.author;
+
+require('./routes/routes.js')(app);
